Restructure
- Moved cmd exec to server
- Added gpio-backend feature to roblib, so client doesn't depend on
  rppal
